Implement dangling<T>/safe_iterator_t as per Ranges TS
Algorithms which return an iterator into a range can potentially
return a dangling iterator if called with an rvalue range.

Previously we static_assert()ed that the passed ranges were
lvalues. Instead, we now implement the Ranges TS solution, which
is to instead return a dangling<Iter> type if called with an rvalue.