Use macros to avoid hard errors in concept definitions
Because type substitution doesn’t short-circuit like real concepts, we
can end up with hard error by, for example, simply saying T& when we’re
handed a T.

This is annoying, and the solution is to use
std::add_lvalue_reference_t etc, which handle void correctly. But
that’s an awful lot of typing, so use macros to make it a bit nicer.