Change namespace from nanorange to nano...
...with an inline namespace ranges.

nanorange:: is a lot to type, and nano:: is better. But some of the range proposals but stuff in std::, and some goes in std::ranges. To make it easier to upgrade to std::[ranges::] when it comes along, we'll let people insert the ::ranges part if they want to.