Fix GCC error with undefined function
GCC doesn't seem to like declaring (but not defining) a specialisation
of a function template with a local type (i.e one defined inside a
function).

This shouldn't be a problem with projected::operator*(), because we
never actually call it, but still.

Since we can't actuall return anything (anything that won't return an
error in some circumstances, anyway), we'll just pretend to throw
something meaningless instead.