Use auto/decltype(auto) returns in view_interface
The spec uses these return types, which is much easier than trying to work out what the return types should be.

This means using enable_if in the template param list rather than on the return type, which is (apparently) slower to compile and plays hell with MSVC is the argument list is too complex, but we'll burn that bridge when we get to it.