Big changes to concept machinery
We had to work hard to avoid hard errors in concept defintions, using checked_foo<T> to guard against cases when trying to form foo<T>  would be an error.

With this commit, things are better: we calculate concepts inside of an enable_if in a function return type, and use SFINAE to ensure that if a concept definiton craps out for any reason, the variable is simply set to false.

Hopefully, this will cure a couple of the nasty edge cases (like void*) in the tests too