(Re)add test_utils.hpp
We're about to make the P0970r1 changes to subrange, which means that safe_iterator_t<subrange> is no longer dangling, and doens't require get_unsafe().

For this reason, we'll go back to using a special make_range() function as we used to, and as the CMCSTL2 tests expect.

Also, get rid of check_equal.hpp while we're at it.