Fix ambiguous calls in 3-legged algos
The Ranges papers don't provide the "three-legged" and "range-and-a-half" versions of various algorithms, but we do so as an extension.

Unfortunately this presents a problem with raw arrays, because they are both ranges (when taken by reference) and iterators (when decayed by value). So in the 3-legged versions we take the param by rvalue reference and check that its decayed type is not a range.

Done for equal, mismatch and transform, might be required for others