Fix subrange on Clang6/C++17 (again)
Clang 6 (or rather libc++) is now using an intrinsic for std::is_constructible in C++17 mode, which causes problems with subrange -- because checking View<T> (which checks Semiregular<T> which checks CopyConstructible<T>) claims that it hits something that is not a constant expression.

I'm not sure whether this is a bug or a genuine problem with NanoRange. As a workaround, we'll just check `detail::view_predicate` instead of `View` for now.