Workaround GCC{5,6} ICE with heap tests
GCC < 7 fails with "confused by earlier errors, bailing out", whihc is just about the least helpful error message imaginable.

I've manage to boil it down to using subrange in the test_utils stuff. I don't know what it is in particular about subrange that's the problem, but we'll jsut using the simple make_range for now as an alternative.