counted_iterator: workaround MSVC not finding iter_move by ADL
counted_iterator has an iter_move friend function which is designed to
be found by ADL in nano::iter_move.

For some reason, MSVC doesn't seem to find this function when it is a
template that is defined in-class; if it is a non-template, or defined
out-of-class, then it seems to work okay.

Unfortunately, defining it out-of-class hits a GCC ICE in the noexcept
condition, so instead we'll just make it a non-template on MSVC only.
This means that we lose the "requires" condition that this
specialisation is only considered when I models InputIterator, but this
doesn't seem too much of a problem.

I ought to check whether this other template friend functions defined
in-class (such as iter_swap here) are correctly picked up by MSVC too.