Implement P0944
This commit attempts to handle contiguous ranges, as best we can.

First, it adds contiguous_iterator_tag, inheriting from random_access_iterator tag. Raw pointers are contiguous; unfortunately, we can't (easily) declare that, for example, vector<T, A>::iterator is contiguous, because we can't deduce T or A in this case. We *might* be able to do this in limited cases (i.e. default allocator); I'll have to give it some thought.

Next, this commit adds the ContiguousIterator concept. This is to spec -- although obviously it's only satisfited by raw pointers at present.

Lastly, this adds a ContiguousRange concept. This is not to spec, because it effectively just requires RandomAccessRange, and that nano::data() returns a poitner to an object. (The spec requires that its iterator type model ContiguousIterator). This deviation from the spec is to allow, e.g. std::vector to model ContiguousRange.