Work around MSVC pointer arithmetic bug
Given

    int* const j
    ptrdiff_t const n

MSVC thinks the type of (j + n) is int* (which is correct), but the type of (n + j) is int* const (GCC, Clang and the RandomAccessIterator requirements all think the latter should be int* as well).

This seems like a compiler bug, but I don't know whether it's one of those things that has been kept for backwards-compat and can be turned off with a compiler switch.

Anyway, we'll work around it by simply removing the type test for (n + j) when using MSVC.