subrange: factor out NotSameAs tests to avoid recursion
This should allow changing the non-member begin() and end() to take a subrange by value rather than by rvalue reference, but the all-in-one fix in #26 caused some CI failures -- so we'll do it in two commits to try to work out what exactly the problem is.

Thanks to @caseycarter for the fix.

Co-Authored-By: Casey Carter <casey@carter.net>