Subrange: Update to P0896R2
This changes subrange's non-member begin() and end() to friend functions taking a subrange by value rather than rvalue reference, per P0896R2.

Unfortuntely, this causes problems with dangling<subrange<I>>, because NotSameAs<dangling, subrange> is true, so the compiler goes ahead to check the second enable_if condition, which winds up in infinite recursion.

The "solution" here is to make safe_subrange<R> ill-formed if R does not model detail::ForwardingRange. This is per the specification in P0896R2, which kills the dangling<> wrapper entirely.

This feels like a bit of a hack, but seems to work with the latest GCC and Clang in local testing. We'll submit it to CI and see how it fares with older compilers.