Replace make_range() with make_subrange() in tests
make_range() was heavily used in the tests, returning a class wrapping an iterator/sentinel pair. This has been replaces with subrange (and, for us, make_subrange) in the proposals, so this commit updates the tests to use it everywhere.

If nothing else, this will be a good test for subrange. Previously when I tried this it had some problems with older GCCs, but there have been some subrange bug fixes since then.

Because subrange is a view, safe_iterator_t<Rng> resolves to just the iterator type, not wrapped in dangling -- which is good, because dangling is going away very soon.