Remove dangling<I> wrapper
With a slightly heavy heart (because I thought it was actually quite a good idea), this commit removes the dangling<I> wrapper and modifies safe_iterator_t to match the P0896R2 spec, so that (effectively) algorithms can only be called on rvalue views, not rvalue containers.

Most of the tests are modifed to hide the rvalue range stuff in #ifdefs, just in case dangling gets added back in at some point.