Add initial implementation of istream_range
This uses the Range-V3 approach of storing the basic_istream pointer and the cached value in the range object itself, rather than the P1035 approach of reusing istream_iterator.

This seems to work okay in local testing -- or at least it passes the Range-V3 istream_range test, which is also added in this commit. A few more thorough tests would probably be a good idea.

One caveat is that this implementation currently requires that the value type of the range is default constructible. Once we add the semiregular wrapper required for Views, we can re-use it here to remove this restriction.