Use our reverse_iterator implementation in rbegin()/rend()
rbegin() and rend() have a fallback case which uses std::reverse_iterator. This commit changes these to use our (hopefully superior) implementation instead.

This required a little bit of reorganisation of headers, because the reverse_iterator implementation needs to use begin() and end(), which were previously defined in the same file as rbegin() and rend(), which led to a circular dependency.

This also requires a "poison pill" for rbegin(std::initializer_list), which otherwise would find the std version via ADL (but strangely, not for rend...).