Special-case std::string_view in nano::begin()/end()
P0896 adds non-member begin() and end() functions which take a string_view by value. Since ranges::begin() does not consider a member begin() when passed an rvalue range, this change allows it to use the non-member version, so that rvalue string_views to be passed directly to algorithms which return iterators -- which is fine, because a string_view does not own its elements and therefore things shouldn't dangle. This is P0970 in action, and it's great.

Unfortunately for us, cracking open namespace std to insert new overloads for begin(string_view) and end(string_view) is strictly forbidden. The only alternative I can think of is to special-case string_view in ranges::begin() and ranges::end() when compiled under C++17, which seems to work okay.