Implement "fast path" buffered stable_partition
The implementation itself is pretty straightforward: we use move_iterators with partition_copy() to move the "false" elements out into the buffer, followed by a move() to put them back into the initial sequence.

This currently uses a regular old std::vector as the buffer; if reserve() throws std::bad_alloc then we swallow it and use the slow algorithm instead.

As far as I know this is technically conforming, but I don't feel very good about it...