Implement our own temporary buffer
So, std::get_temporary_buffer() is deprecated in C++17, and set to be removed in C++20. Yet three algorithms require trying to get temporary storage, falling back to a slower algorithm if that's not possible, so we need to have some way to do that.

The easiest way is simply to use std::vector with the default allocator,  and catch std::bad_alloc. However, for some compilers the very presence of a try/catch block inhibits optimisation, and other people will want to compile with exceptions disabled entirely.

So instead, we'll reserve our own temporary storage using operator new(std::nothrow), and check whether this returns nullptr. If it succeeds (and let's face it, it probably will), then we'll wrap it in an allocator and use it with std::vector.

Whether or not this actually leads to better codegen, I have no idea. I also have no idea whether I've actually written a valid allocator, although it seems to work...