Rotate: add some special cases
This commit adds three special-cases to rotate(), to try to tackle #42

Firstly, if we are rotating two equally-sized ranges then we can use `swap_ranges` instead -- this is always a perf win, but (alarmingly) very much more so if we use the three-iterator form. The extra comparison of the four-element form makes things aroung *10x* slower. (Which calls into question the wisdom of deprecating the three-legged versions of these algorithms, but that's a topic for another time.)

Secondly, if the "middle" iterator is one position past "first", then we stash the first value, move all the other elements back one place, and then finally move the element back into the end of the sequence. For trivial types, this is a perf win, but alas still not as fast as libc++ or libstdc++ manage. I suspect this is because these implementation can optimise `move` to `memmove`,  which we cannot do due to the `constexpr` requirements. For non-trivial types (tested with `std::string`) this is very much slower than the naive algorithm when using libstdc++. I'm not sure why that is, but it's enough to convince me that we should use it only for types which are trivially move-assignable (which is what libc++ does).

Finally, if middle is one position before the end of the sequence, and we are using bidirectional iterators, we perform the reverse of the above, using `move_backward`