Replace std::enable_if_t with our own custom version
The idea here is that instantiating class templates is slow, but aliases are fast. As specified, std::enable_if_t is an alias for std::enable_if<Bool, Type>::type. So every time we use it, we are (in theory) instantiating a new specialisation of enable_if for each type.

Instead, if we define enable_if as

    template <bool> {
    struct enable_if {
        template <typename T>
        using type = T;
    };

    template <>
    struct enable_if<false> {};

    template <bool B, typename T = void>
    using enable_if_t = typename enable_if<B>::template type<T>;

then in theory we only every instantiate two versions of the enable_if class template for true and false, and use the template alias (if present) to get the actual type. So, much faster, right?

Alas, no. This turns out to actually be a slightly bit slower with GCC. And we don't want to use it with Clang, as that special-cases std::enable_if to give lovely error messages in newer versions.

Boo.