[gcc9] Fix misbehaving tests with GCC9
These tests are copied from Range-V3/CMCSTL2 and work with previous versions, but mysteriously fail with GCC9.

I'll be honest, I have no real idea what the problem is (something to do with designated initialisers? A regression in the new version?) but copying the Range-V3 solution of explicit captures rather than a capture-default seems to fix it, so... 🤷🏻‍♂️