Require C++17
NanoRange was originally concieved as a "small" implementation of the Ranges TS, or at least small compared to Range-V3. Over time it has grown substantially, and quite a lot of complexity and workarounds have been added to support older C++14 compilers (in particular GCC5).

By requiring C++17, we can keep NanoRange closer in line with what it was intended to be, by removing workarounds for older compilers. In addition, being able to use C++17 features (in particular if constexpr and constexpr lambdas) means that we can (hopefully) try to keep our compile times down.

I imagine the intersection of users who want the latest cutting-edge C++ library features, but are stuck with older compilers, is rather small. For those that do exist, Range-V3 still supports C++14 for now.