Avoid weird GCC9.1 static_assert failure
For some reason GCC9.1 doesn't think reverse_view(int[10]) is a RandomAccessRange -- even though previous versions of GCC do, and so does Clang and MSVC.

In fact, it doesn't seem to think it's even an InputRange, despite the fact that Range<T> and InputIterator<iterator_t<T>> both pass.

My only guess at the moment is some sort of weird compiler bug. As such, we'll comment out this assertion for GCC9.1 (only).