Initial, incomplete attempt at piping views
The idea is that we mark our view::xxx objects as "range adaptor closure objects" (RACOs) via a trait, and have a proxy object which itself is also a RACO for when the user "pipes" two of them together.

Then, we have a generic operator|() (in our detail namespace, found via ADL) which is used when the LHS is a (non-RACO) ViewableRange and the RHS is a RACO.

This does not (yet) handle "partial application", i.e. saying something like vec | view::transform(func), mostly because we don't yet have any views that take multiple arguments. But it seems to work so far.

This approach requires explicitly opting in to being a RACO, so users can't easily add their own pipeable views (unless they specialise the private is_raco trait), but my reading of the standardese suggests this is okay.