Avoid MSVC bug in iota_view test
In MSVC, using iota_view with a type that doesn't have a constexpr operator== (and probably !=) leads to an error about "a call to a function which can never be constexpr" during constraint deduction.

This appears to be a compiler bug which was fixed in the very latest (at the time of writing, two days old) version 16.2 -- or at least, it works fine on my system with that compiler.

I'm not sure whether this is actually a major problem -- I'd imagine 99% of uses of iota_view will be with built-in integer types -- but it's probably worth investigating further. But for now, we'll just disable the offending test.