Travis: drop xcode9.3, add xcode11
Whichever "real" version of Clang XCode 9.3 corresponds to, it appears to implement C++17 just fine. Unfortunately Apple in their wisdom decided not to ship the corresponding headers for <any>, <variant> and, in particular, <optional>

Since we're using the latter in the new views, now is the time to drop compilers that don't support C++17 properly.

To make up for it, and because I just love watching Travis compile things, we'll add the latest XCode11 image to see what Apple have decided to break recently.