Allow piping to view::filter(pred)
We use a lambda to capture the given predicate and then forward (well, move) that into the constructor call later. In order to make the lambda a "range adaptor closure object", we wrap it in a `detail::rao_proxy` object.

This seems far too simple to actually work, but surprisingly, it does -- at least for the use cases in the filter tests.