filter_view: GCC7/8/correctness(?) fix
This op== is a hidden friend of filter_view::sentinel, and so should have no access rights to filter_view::iterator, and in particular its current_ private member.

This code fails on GCC 7 and 8, but passes with GCC 9 and all tested versions of Clang. Unless I'm missing something, I'm pretty sure this should be rejected everywhere.

Anyway, the fix is simple enough. And we'll also fix the slighly embarassing typo of using = instead of ==...