Move some definitions around to try to appease MSVC
MSVC < 16.1 complains that it can't find an appropriate version of operator| when the RHS is a rao_proxy object.

This appears to be another MSVC bug because it works fine with 16.1 and GCC/Clang.

As a first try, let's attempt to move round some definitions so that MSVC knows that is_raco<rao_proxy> is true the first time it sees either of the op| overloads. It might work...