MSVC "piping" fix, take 2
In MSVC 15.9 and 16.0, the compiler complains that it cannot find and appropriate overload of operator| for an iota_view on the LHS and a raco_proxy on the RHS.

There are a couple of things that could be going wrong here: firstly it's simply not finding them via ADL, or secondly there's a problem with constraints and neither overload is seen as a candidate.

As a second attempt at a "simple" fix, we'll try changing the way we write the contraints for these overloads from being part of the template param list to the return type, and see if that makes a difference.