MSVC pipe fix, take 3
So after some investigation, it turns out the problem is using a lambda as  template argument: for some reason it seems to prevent the compiler from successfully finding the op| definitions.

As a rather heavy-handed workaround, and because I can't think of a better way to do it right now, we'll use std::bind on MSVC instead, which seems to work correctly.

Because this is a frankly rubbish solution, but one we're going to have to use in a few places, we'll hide it behind a macro. I'm not quite sure of the target version yet: in my local testing, VS16.2 (=1922) appears to suffer from the bug, but in the CI runs VS16.1 (=1921) seemed to work correctly. For the moment we'll just set it to 1921 and see if we get any complaints.