Update algo return values to handle dangling
Back when dangling used to be a thing, algorithms that returned multiple iterators used std::pair or std::tuple (in our implementation, anyway). These have converting constructors which handled converting a pair<iterator, iterator> into a pair<dangling, dangling> in the return value.

Nowadays the return types are simple aggregates instead, and the conversion can't happen automatically, so we need to do it manually. Annoying, but not the end of the world.