Re-enable some algo tests on rvalue ranges
A long time ago when  removing dangling v1, I commented out a lot of rvalue range tests rather than deleting them entirely, in the hope that one day dangling would make its triumphant return.

Today, that day has arrived.

But alas, dangling is not the class it once was, for it no longer allows calling get_unsafe(). So many of the tests needed to be changed anyway. Ah well.