Avoid broken rvalue C arrays with MSVC
For some reason MSVC refuses to believe that a moved C array is an rvalue. A bit of experimenting with Compiler Explorer suggests this was partially fixed in 19.22, but universal references still seem to deduce the wrong type even with the latest version. I have, for a change, reported the bug rather than just whinging about it in a commit message: https://developercommunity.visualstudio.com/content/problem/700039/incorrect-behaviour-with-rvalue-c-arrays-and-unive.html

Anyway, this caused a lot of failures with the dangling tests with MSVC.  As a workaround, we'll use std::array instead.