Work around really odd GCC9 bug
Given X = reverse_view<int(&)[10]>, GCC9 for some reason gets very confused and fails the static_assert if you ask it to check RandomAccessRange<X> *after* checking View<X>. In fact, it won't even admit that X is an InputRange -- but if you give it an identical definition of the InputRange concept locally then it passes.

If you remove the View check, then it works. If you add a ForwardRange check *before* the View check -- as in this commit -- then it works (but, curiously, a prior InputRange check is not sufficient to make it pass).

This is all very strange and I have no idea what's going on -- in fact, it's all a little concerning. Fortunately, everything works okay in GCC7 and 8, Clang and MSVC, so at this point I can only assume it's a really odd compiler bug.