Ugly, ugly workaround for GCC9 drop_view test failures
Mysterious random contraint failures with GCC9 strike again. I still haven't gotten to the bottom of what causes them (or even the top, really) and this time I wasn't able to make the failures go away by reordering static_asserts. I'm starting to get very concerned about it.

As an ugly workaround, we'll disable the bidirectional_range assertion when the compiler is GCC9. By way of compensation, we'll add a contraint to the deduction guide.