Fix broken piping on MSVC (again)
Thanks to @QuellaZhang for pointing out that the problem with our lambda piping approach is not inheriting from a lambda-as-template-arg (as I thought), but actually the complex explicit return type of the lambda itself.

By using a deduced return type when the compiler is MSVC we can use the same nifty approach to piping everywhere. Using a deduced return type like this doesn't actually lose us very much either.

This issue with return types is fixed with MSVC's new lambda processor, which is enabled with /std:c++latest or /experimental:newLambdaProcessor. AFAIK there is no way to detect at compile-time whether this is in use, so for now we'll unconditionally use the workaround if the compiler is MSVC. When the new processor becomes the default, we can add a version check to the workaround macro.

This commit hopefully finally fixes #54 properly