Use the new MSVC piping workaround
Sadly, this still failed in one of the take_view tests where we did rng | filter(f) | take(n). As yet another workaround, we'll use a helper alias to explictly specify the return type of the lambda. I suspect the fun times with MSVC piping haven't finished yet. *sigh*

(It all works okay with the new lambda processor though, so there is hope for the future.)