take_while_view: avoid bug in old MSVC versions
Before 16.2, MSVC would incorrectly diagnose that a call to a friend function could never be constexpr. The workaround is to make the functions templates instead. But since instantiating a function template is (probably) infintesimially slower, we'll avoid doing it if we don't have to.

Because, y'know, we're all about fast compile times.