Use a different workaround for join_view::iterator<true>
In https://github.com/ericniebler/stl2/issues/604, the proposed solution is to mark the inner_ member as mutable.

This works fine, except in a constexpr context, where both Clang and GCC complain about a read of a mutable member.

This being the case, we'll use an alternative workaround suggested by Casey in that thread.

(Who knew you couldn't read from a mutable member in a constexpr function?)