Avoid having to define projected::operator*()
I've never quite gotten to the bottom of why GCC requires this to be defined when it's never called, but it's bothered me for a long time.

Anyway, we can work around it (and potentially save ourselves a few template instantiations in contraint checks) by special-casing projected<Iter, identity> to be just Iter.