Make istream_view::iterator copyable (again)
While the concepts (and the istream_view spec) have been updated for move-only iterators, the views have not (yet).

Since it's handy to be able to apply adaptors to an istream_view, we'll comment out the move constructor definitions for now.

(Oh, and also work around yet another mysterious GCC constraint failure, grrr....)