Use structured bindings in split_view
I was under the mistaken impression that you couldn't use structured bindings in a constexpr function. In fact you can, you just can't declare the bindng itself to be constexpr.

I'm still not quite sure why the spec says to use subrange rather than just calling begin() and end() directly on the pattern object, but at least doing it this way is a bit tidier.