Update next_/prev_permutation result type
Apparently the result type of prev_permutation and next_permutation was changed, so we'll use the new version. I'm pretty sure it needs [[no_unique_address]] and a conversion operator like the other _results, but it doesn't have one at the moment so never mind.

The tests just check the result of found, should probably update them to check the returned iterator as well...