Use borrowed_range instead of forwarding_range
This is the first of what will probably be a large set of changes to get (as close as we can) to C++20 conformance.

First up is removing forwarding_range and replacing it with a new borrowed_range opt-in trait. This is a large and scary change and affects some of the most fundemental concepts and functions in the library, including the definitions of ranges::begin() and ranges::end(), which can now call member begin() on rvalues provided they specialise enable_borrowed_range.