[skip appveyor] Try moving to Xenial for Travis builds
Apparently Clang 9 packages don't exist in the LLVM repo for Ubuntu Trusty. They do exist for the newer Xenial release though, which is now the Travis default image. We'll try moving everything to Xenial and see what breaks: if the answer is "stuff", we'll revert this and use Xenial only for GCC 9 and Clang 9