Unconditionally use invocable workaround on AppleClang
Not unexpectedly, simply attempting to detect whether __clang_major__ is less than 7 does not work with AppleClang, which inexplicably (and maddeningly) defines it to be the XCode version number instead.

Since I really can't be bothered to try and work around Apple's idiocy on this matter, we'll just unconditonally use the workaround if we detect the (completely undocumented, natch) __apple_build_version__ symbol.