Update view concept definition to match C++20 spec
If you're going to write a blog post about something, you should probably make sure your own library implements it correctly, right?

This commit removes the heuristic that assumes something is a view if begin() and cbegin() return the same type, which is rather wonderful and means we no longer have to do the ugly special-casing of lots of standard library types (other than string_view, which I can live with).